Split profiles into its own file
